web/requirements.txt
APScheduler
Flask_HTTPAuth
Flask_Table
Flask
matplotlib
numpy
pandas
psycopg2-binary
requests
scipy
uWSGI
